Three commits across two Jira tickets in one PR, per project decision:
1. PTIMER-209 — iOS fractional ND presets
2. PTIMER-209 — Android parity
3. PTIMER-220 — Android relaunch crash

## PTIMER-209 — Fractional ND presets (iOS + Android)

Adds three permanent commercial fixed-ND presets to the ND wheel as
configured one-decimal stop values, so high-density filters that do not
map to a whole stop calculate correctly instead of rounding:

| Stops | Optical density | Filter factor |
|-------|-----------------|---------------|
| `6.6` | `OD 2.0` | `ND100` |
| `7.6` | `OD 2.3` | `ND200` |
| `16.6` | `OD 5.0` | `ND100k` |

`6.6` etc. are the app's chosen one-decimal values, not the exact `log2`
of the marketed factor (`log2(100) ≈ 6.644`).

- **Stops stay canonical.** Each preset feeds the exposure engine as its
configured fractional stop value (`base × 2^stops`, no snap), never
rounded to a neighbouring whole stop. The presets are spliced into the
single-source ND ladder in numeric order (`… 6, 6.6, 7, 7.6, 8, … 16,
16.6, 17 …`), present regardless of the active notation.
- **Notation round-trips.** Selection tracks the canonical stop value
only, so Stops / OD / ND switching returns to the same wheel row. Moving
off `16.6` just selects the neighbouring stop; there is no separate
`ND100k` identity.
- **Display.** The formatter renders these off-grid values as a decimal
(`6.6`) and maps their marketed factor labels (`ND100` / `ND200` /
`ND100k`) rather than `2^stops` (which would read `ND97` / `ND194` /
`ND97K`). OD is `stops × 0.3`.
- **Fixed product set is an invariant.** The three products are the only
non-integer values the picker exposes and the only off-grid values
eligible for commercial labels and exact persistence. Formatter,
writers, and restore route through `commercialNDPresetStop`; the
scale-flip sanitizer canonicalizes against the same ladder. A near-match
normalizes to canonical; an unsupported off-grid value is ignored. The
reserved third-stop `NDStep` path remains for a future custom-ND
workflow.
- **Persistence.** An additive optional exact-stops field is used on
both platforms (iOS `ndStopsExact`, Android `ndStops`) and is preferred
on restore only when it matches a preset. There is no schema-version
bump; existing stored data remains backward-compatible. Android
decouples wheel position from canonical stop value and widens
`ShootingCalculator` to `Double`.
- Specs updated (Calculator §2.2 / §1.4, UI ND picker). Patch version
`0.7.10 → 0.7.11` on both platforms.

No PTIMER-199 / ND stack mode, no arbitrary-decimal input, no new filter
inventory, and no new product data model.

## PTIMER-220 — Android app can crash when relaunched with a running
timer

Relaunching the Android app while a timer was running could crash during
startup with `ForegroundServiceDidNotStartInTimeException`. The change
is isolated to `TimerForegroundService.kt`; no ND calculation or picker
code is involved.

## Verification

All automated results below are local; no CI is wired to this
repository.

iOS:
- `swift test --package-path ios/PTimerKit` — 1404 passed
- `xcodebuild … -testPlan PTimer test` (iPhone 17) — TEST SUCCEEDED
- `cd ios && swiftlint lint` — clean on changed files

Android:
- `./gradlew :core:test :app:testDebugUnitTest :app:assembleDebug
:app:lintDebug` — BUILD SUCCESSFUL, including ND
ladder/formatter/controller tests and real `SlotSessionCodec` JSON
round-trip cases
- Emulator (Pixel 10, API 36): `ND100k` renders between `ND64K` and
`ND128K`, and `1/30 + 16.6 stops` computes `3311s / 55m 11s`; Stops / OD
/ ND switching works
- Relaunch with a running timer no longer crashes in the reproduced
emulator sequence, and the ongoing timer notification remains available

Cross-platform capture check:
- iOS and Android both show `1/30`, `16.6 stops`, adjusted exposure
`3311s / 55m 11s`, and FP4 Plus corrected exposure `27238s / about 7h
33m`

## User test steps

PTIMER-209, both platforms:
1. Open the ND picker on the shooting screen.
2. In Stops, confirm `6.6`, `7.6`, and `16.6` appear between 6/7, 7/8,
and 16/17.
3. In ND, select `ND100`, then switch to Stops and confirm `6.6` is
selected.
4. Select `ND200`, then switch to Stops and confirm `7.6` is selected.
5. Select `ND100k`, switch to Stops and confirm `16.6`; switch to OD and
confirm `5.0`; switch back to ND and confirm `ND100k`.
6. From `16.6`, move the wheel to `16` or `17` and confirm the
calculated result changes immediately.
7. Select an existing integer preset such as ND1000 and confirm its
previous behavior is unchanged.
8. Relaunch the app, or switch camera slots and back, and confirm the
fractional preset restores exactly.

PTIMER-220, Android:
9. Start a timer.
10. Terminate and reopen the app using the same relaunch sequence that
reproduced the crash while the timer is still running.
11. Confirm the app opens without crashing, the running timer remains
available, and the ongoing notification remains present.
12. Stop the timer and confirm normal timer stop behavior still works.

+++ android/app/src/main/kotlin/com/sangwook/ptimer/app/calc/ShootingCalculator.kt
@@ -26,11 +26,11 @@ class ShootingCalculator(
 ) {
     private val ladder = ExposureScale.oneThirdStop.shutterSteps
 
-    fun result(shutterIndex: Int, ndStops: Int, profile: ReciprocityProfile?): ShootingResult {
+    fun result(shutterIndex: Int, ndStops: Double, profile: ReciprocityProfile?): ShootingResult {
         val base = ladder[shutterIndex.coerceIn(ladder.indices)].seconds
         val adjusted = exposure.calculate(
             baseShutterSeconds = base,
-            ndStep = NDStep(ndStops.toDouble()),
+            ndStep = NDStep(ndStops),
             scaleMode = ExposureScaleMode.ONE_THIRD_STOP,
         )
